requirements.txt
pymongo==4.6.1
pandas==2.2.0
streamlit==1.31.0
python-dotenv==1.0.0
plotly==5.18.0
numpy==1.26.3
faker==22.6.0 